Improve C++ modules

+++ model/model.py
@@ -46,7 +46,7 @@ class Model:
         """Base class representing the parameters."""
 
         max_dimension: int = 0
-        network_size: int = 0
+        network_size: float = 0.
 
     def set_relevant_parameters_from_data_set(self, data_set: DataSet) -> None:
         """Set the model parameters based ona a data set."""
